TileMap 1
This is a rendition of the first tilemap which can be used for either the black bears or brown bears.

Please note that berries have not been populated yet, but will be added layer in a top-most layer.

Layers Branch_1 and Leaves_1 can be rendered by code at once, as well as layers Branch_2 and Leaves_2. When rendered, we can spawn berries for the player to eat on the branches/leaves spawned.